requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==2.2.13
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-tempus-dominus==5.1.2.17
gunicorn==20.0.0
idna==3.3
Pillow==9.1.1
psycopg2-binary==2.9.3
python-dotenv==0.10.3
pytz==2019.3
requests==2.27.1
sqlparse==0.3.0
stripe==2.38.0
urllib3==1.26.9
whitenoise==4.1.4
